Move AssayDefinitionModel.ts to a separate file
