+++ packages/components/src/internal/AssayDefinitionModel.spec.ts
@@ -1,28 +1,6 @@
-/*
- * Copyright (c) 2019 LabKey Corporation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-import { fromJS, List, OrderedMap } from 'immutable';
-
-import assayDefJSON from '../../../../test/data/assayDefinitionModel.json';
-import assayDefNoSampleIdJSON from '../../../../test/data/assayDefinitionModelNoSampleId.json';
-
-import {
-    AssayDefinitionModel,
-    AssayDomainTypes,
-    SchemaQuery,
-} from '../../../..';
+import { AssayDefinitionModel, AssayDomainTypes, SchemaQuery } from '../index';
+import assayDefJSON from '../test/data/assayDefinitionModel.json';
+import assayDefNoSampleIdJSON from '../test/data/assayDefinitionModelNoSampleId.json';
 
 describe('AssayDefinitionModel', () => {
     test('with getSampleColumn()', () => {
@@ -83,4 +61,3 @@ describe('AssayDefinitionModel', () => {
         expect(dataColumns.has('date')).toBeTruthy();
     });
 });
-
