+++ packages/components/src/internal/components/assay/AssayReimportHeader.spec.tsx
@@ -4,7 +4,7 @@ import renderer from 'react-test-renderer';
 import { fromJS } from 'immutable';
 
 import assayDefJSON from '../../../test/data/assayDefinitionModel.json';
-import { AssayDefinitionModel } from '../base/models/model';
+import { AssayDefinitionModel } from '../../AssayDefinitionModel';
 
 import { AssayReimportHeader } from './AssayReimportHeader';
