+++ packages/components/src/internal/components/assay/RunDataPanel.spec.tsx
@@ -7,7 +7,7 @@ import { gridInit } from '../../actions';
 import { withFormSteps, WithFormStepsProps } from '../forms/FormStep';
 import { ASSAY_WIZARD_MODEL } from '../../../test/data/constants';
 import { initUnitTestMocks } from '../../testHelpers';
-import { AssayUploadTabs } from '../base/models/model';
+import { AssayUploadTabs } from '../../AssayDefinitionModel';
 
 import { RunDataPanel } from './RunDataPanel';
